API: Create an APi for Payment Service

+++ backend/src/main/java/ecommerce/service/MicrofinanceService.java
@@ -51,7 +51,7 @@ public class MicrofinanceService {
         
     }
 
-    public String createFeeRequest(String microfinanceUsername, float amount, String user) {
+    public String createFeeRequest(String microfinanceUsername, double amount, String user) {
         String url = baseUrl + "/transactions/service/ecommerce/create-fee-request/";
 
         String jsonBody = String.format(
